requirements.txt
Flask~=3.1.0
PyPDF2==3.0.1
openai
python-dotenv==1.0.0